# booking-service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Slot scheduling and booking system for KBTU counseling platform.",
        "title": "KBTU Psychologist Booking Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/psychologist/reviews": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Psychologist views their ratings and written reviews. Student identities and exact dates are anonymized to protect privacy.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "psychologist-slots"
                ],
                "summary": "Get anonymous reviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AnonymousReviewResponse"
                            }
                        }
                    }
                }
            }
        },
        "/psychologist/slots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all slots for the logged-in psychologist, including student details for booked slots.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "psychologist-slots"
                ],
                "summary": "Get psychologist's own schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter period: 'day', 'week', 'month' (default: day)",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date for filter (YYYY-MM-DD), default: today",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PsychologistScheduleResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Psychologist generates multiple slots based on a weekly schedule pattern between start_date and end_date.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "psychologist-slots"
                ],
                "summary": "Create schedule slots for a psychologist",
                "parameters": [
                    {
                        "description": "Schedule configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateScheduleInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleCreatedResponse"
                        }
                    },
                    "400": {
                        "description": "validation error or no slots created",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "only psychologists can create slots",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/psychologist/slots/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows a psychologist to remove an available slot from their schedule. Fails if the slot is already booked.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "psychologist-slots"
                ],
                "summary": "Delete an unbooked slot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not authorized or not the owner",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Slot not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Cannot delete a booked slot",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/psychologist/slots/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Psychologist cancels a session. Frees the slot and triggers a cancellation email to the student.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "psychologist-slots"
                ],
                "summary": "Psychologist cancels a booked appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "Not authorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Slot not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Slot is not booked",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/psychologist/slots/{id}/notes": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Psychologist writes private notes after a session. Hidden from the student.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "psychologist-slots"
                ],
                "summary": "Add private notes to a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The private notes",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddNoteInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/psychologist/statistics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns meaningful KPIs: Load, Ratings, Trends, and efficiency.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "psychologist-slots"
                ],
                "summary": "Get psychologist dashboard statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PsychologistStats"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AnonymousReviewResponse": {
            "type": "object",
            "properties": {
                "month_year": {
                    "description": "e.g., \"February 2026\"",
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "review": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid start_date"
                }
            }
        },
...
        "models.PsychologistScheduleResponse": {
            "type": "object",
            "properties": {
                "booking_type": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "psychologist_id": {
                    "type": "string"
                },
                "questionnaire_answers": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "student_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}